Reset http->pos
Reset the http reader position to the begenning when a line is fully
read. Shouldn't be usefull in most cases, but cause error (BAD_JSON) if
the socket isn't closed between two requests and a request is smaller
than the previous one.

Fix a problem I had while connecting to APE using a normal socket
client.